taal

+++ hooks/voorpagina.php
@@ -3,11 +3,11 @@
 if (!function_exists('ag_vp_print_nieuws_hook')) : function ag_vp_print_nieuws_hook()
 	{
 
-		echo "<section class='vp-nieuws verpakking verpakking-klein'>
-	<h2>Nieuws</h2>";
+		echo "<section class='vp-nieuws verpakking'>
+	<h2>" . ucfirst(\agitatie\taal\streng('nieuws')) . "</h2>";
 
 		$vp_posts = new WP_Query(array(
-			'posts_per_page' => 6
+			'posts_per_page' => 5
 		));
 
 		echo "<div class='art-lijst'>";
@@ -36,7 +36,7 @@ if (!function_exists('ag_vp_print_nieuws_hook')) : function ag_vp_print_nieuws_h
 
 		$k = new Ag_knop(array(
 			'link' 		=> get_post_type_archive_link('post'),
-			'tekst' 	=> 'alle berichten',
+			'tekst' 	=> ucfirst(\agitatie\taal\streng('alle')) . ' ' . \agitatie\taal\streng('posts'),
 			'class'		=> 'in-wit'
 		));
 		$k->print();
@@ -44,44 +44,68 @@ if (!function_exists('ag_vp_print_nieuws_hook')) : function ag_vp_print_nieuws_h
 		echo "</footer>";
 
 		echo "</section>";
+
+		echo "<section class='verpakking verpakking-klein marginveld'>";
+		echo "<div class='agenda'>
+		<h2>Agenda</h2>";
+
+
+
+		//$afm = ag_agenda_filter_ctrl();
+
+		if (have_posts()) :
+
+			$agenda = new Ag_agenda(array(
+				'aantal' => 10,
+				'omgeving' => 'pagina'
+			));
+
+			$agenda->print();
+		endif;
+
+		echo "</div>";
+		echo "</section>";
 	}
 endif;
 
 add_action('voorpagina_na_tekst_action', 'ag_vp_print_nieuws_hook', 20);
 
 
-function ag_vp_print_menu()
-{
+if (!function_exists('ag_vp_print_menu')) : function ag_vp_print_menu()
+	{
 
-	$locaties = get_nav_menu_locations();
+		$locaties = get_nav_menu_locations();
 
-	if (array_key_exists('voorpagina', $locaties)) {
+		if (array_key_exists('voorpagina', $locaties)) {
 
-		$menu = wp_get_nav_menu_object($locaties['voorpagina']);
-		$menu_stukken = wp_get_nav_menu_items($menu->term_id);
+			$menu = wp_get_nav_menu_object($locaties['voorpagina']);
+			$menu_stukken = wp_get_nav_menu_items($menu->term_id);
 
-		if ($menu_stukken and count($menu_stukken)) :
-			echo "<section class='vp-menu verpakking verpakking-veel paddingveld marginveld veel normale-padding achtergrond-zijkleur'>";
-			echo "<h2 class='tekst-wittig geen-margin-top lineheight-fix' >Lees meer over...</h2>";
+			if ($menu_stukken and count($menu_stukken)) :
 
-			echo "<nav class='knoppendoos groot'>";
-			foreach ($menu_stukken as $menu_stuk) {
-				$k = new Ag_knop(array(
-					'link' 		=> $menu_stuk->url,
-					'tekst'		=> $menu_stuk->title,
-					'class'		=> 'in-wit'
-				));
-				$k->print();
-			}
-			echo "</nav>"; //Ag_knoppendoos
 
-			echo "</section>";
 
-		endif;
+				echo "<section class='vp-menu verpakking paddingveld marginveld veel normale-padding achtergrond-zijkleur'>";
+				echo "<h2 class='tekst-wittig geen-margin-top lineheight-fix' >" . \agitatie\taal\streng('Zie ook') . "</h2>";
+
+				echo "<nav class='knoppendoos groot'>";
+				foreach ($menu_stukken as $menu_stuk) {
+					$k = new Ag_knop(array(
+						'link' 		=> $menu_stuk->url,
+						'tekst'		=> $menu_stuk->title,
+						'class'		=> 'in-wit'
+					));
+					$k->print();
+				}
+				echo "</nav>"; //Ag_knoppendoos
+
+				echo "</section>";
+
+			endif;
+		}
 	}
-}
-
-add_action('voorpagina_na_tekst_action', 'ag_vp_print_menu', 10);
+endif; // if ag_vp_print_menu exists
+add_action('voorpagina_na_tekst_action', 'ag_vp_print_menu', 15);
